+++ lib/pushmi_pullyu/cli.rb
@@ -199,6 +199,8 @@ class PushmiPullyu::CLI
   def run_preservation_cycle
     begin
       entity = queue.wait_next_item
+      PushmiPullyu::Logging.log_preservation_attempt(entity,
+                                                     queue.get_entity_ingestion_attempt(entity))
       return unless entity && entity[:type].present? && entity[:uuid].present?
     rescue StandardError => e
       log_exception(e)
@@ -212,7 +214,7 @@ class PushmiPullyu::CLI
         # Push tarred AIP to swift API
         deposited_file = swift.deposit_file(aip_filename, options[:swift][:container])
         # Log successful preservation event to the log files
-        PushmiPullyu::Logging.log_preservation_event(deposited_file, aip_directory)
+        PushmiPullyu::Logging.log_preservation_success(deposited_file, aip_directory)
       end
     # An EntityInvalid expection means there is a problem with the entity information format so there is no point in
     # readding it to the queue as it will always fail
@@ -221,7 +223,9 @@ class PushmiPullyu::CLI
       log_exception(e)
       begin
         queue.add_entity_in_timeframe(entity)
+        PushmiPullyu::Logging.log_preservation_fail_and_retry(entity, queue.get_entity_ingestion_attempt(entity), e)
       rescue PushmiPullyu::PreservationQueue::MaxDepositAttemptsReached => e
+        PushmiPullyu::Logging.log_preservation_failure(entity, queue.get_entity_ingestion_attempt(entity), e)
         log_exception(e)
       end
 
@@ -244,7 +248,11 @@ class PushmiPullyu::CLI
 
   def setup_log
     if options[:daemonize]
-      PushmiPullyu::Logging.initialize_logger(PushmiPullyu.application_log_file)
+      PushmiPullyu::Logging.initialize_loggers(
+        log_target: PushmiPullyu.application_log_file,
+        events_target: "#{PushmiPullyu.options[:logdir]}/preservation_events.log",
+        json_target: "#{PushmiPullyu.options[:logdir]}/preservation_events.json"
+      )
     else
       logger.formatter = PushmiPullyu::Logging::SimpleFormatter.new
     end