+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and releases in PushmiPullyu adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+ - Add more logging information [#433](https://github.com/ualbertalib/pushmi_pullyu/issues/433)
 
 - Add logic to perform authentication against the V3 Auth protocol